fix: make mean's output a scalar instead of a 2d array
To ensure `lower` and `upper` are a scalar, `x` must be a 1d array.
Otherwise, if `x` is a 3d array, `lower` and `upper` will be 2d array, which are not what
we want in our context. However, we tend to have x as a `[1, len(x), 1]`. In this case,
we need to flatten `x` to 1d array to get the correct `lower` and `upper`.